fiscalCodeSchema to common and ulid fix
