+++ packages/io-messages-common/src/domain/fiscal-code.ts
@@ -1,0 +1,11 @@
+import { z } from "zod";
+
+export const fiscalCodeSchema = z
+  .string()
+  .regex(
+    /^[A-Z]{6}[0-9LMNPQRSTUV]{2}[ABCDEHLMPRST][0-9LMNPQRSTUV]{2}[A-Z][0-9LMNPQRSTUV]{3}[A-Z]$/,
+    "Must be a valid Italian fiscal code",
+  )
+  .brand("FiscalCode");
+
+export type FiscalCode = z.infer<typeof fiscalCodeSchema>;
